new added

+++ InventoryPractical/Models/Inventory.cs
@@ -3,8 +3,8 @@
     public class Inventory
     {
         public string ProductCode { get; set; }
-        public int EventType { get; set; }
-        public int Quantity { get; set; }
+        public double EventType { get; set; }
+        public double Quantity { get; set; }
         public double Price { get; set; }
         public DateTime Date { get; set; }
     }
